feat(cart): extract action class for RetrievingCarProducts to use in cart and checkout
